added pages (previously tabs) added locations
