+++ static/gdpr/ita/gdpr.md
@@ -1,0 +1,50 @@
+# Informativa sul trattamento dei dati personali
+
+## 1. Titolare del trattamento dei dati
+Il titolare del trattamento è ADMIN. 
+Per qualsiasi domanda o richiesta è possibile contattarci all’indirizzo email: marco.riezzo@yahoo.it.
+
+## 2. Dati personali raccolti
+L’applicazione raccoglie le seguenti informazioni personali:
+
+* Street Name (via/strada)
+* City (città)
+* House Number (numero civico)
+* Country (nazione)
+
+## 3. Finalità del trattamento
+I dati raccolti saranno trattati esclusivamente per i seguenti scopi:
+
+1. Facilitare l’organizzazione di tavoli da gioco e condividere l’indirizzo con i partecipanti interessati.
+2. Invio di un messaggio tramite Telegram contenente l’indirizzo del luogo di gioco nel canale della community di cui i giocatori fanno parte.
+
+## 4. Base giuridica del trattamento
+Il trattamento dei dati è basato sul consenso esplicito fornito dall’utente al momento della creazione di una proposta di gioco.
+
+## 5. Modalità di trattamento e sicurezza
+I dati sono conservati in un database protetto e accessibili esclusivamente agli utenti autorizzati secondo le seguenti modalità:
+
+* L’indirizzo è visibile agli altri utenti solo se associato a un tavolo da gioco attivo.
+* I dati non verranno mai utilizzati per scopi pubblicitari o di marketing.
+* Misure di sicurezza come crittografia e accesso limitato sono implementate per proteggere i dati.
+
+## 6. Condivisione dei dati
+L’indirizzo sarà condiviso solo nei seguenti contesti:
+
+1. Tra gli utenti partecipanti al tavolo di gioco associato.
+2. Tramite un messaggio Telegram inviato nel canale della community di appartenenza.
+
+## 7. Conservazione dei dati
+I dati saranno conservati nel database in forma definitiva ma l'utente potrà in qualunque momento cancellare e rimuovere un indirizzo. 
+Una volta che il tavolo non sarà più attivo, comunque, l'indirizzo non sarà più visibile in app.
+
+## 8. Diritti dell’utente
+L’utente ha diritto di:
+
+* Accedere ai propri dati.
+* Rettificare i dati inesatti.
+* Cancellare i dati inseriti.
+* Limitare o opporsi al trattamento dei dati.
+
+## 9. Aggiornamenti alla presente informativa
+Ci riserviamo il diritto di aggiornare la presente informativa per adeguarla a eventuali modifiche normative o funzionali. Gli utenti saranno informati di eventuali modifiche rilevanti.
